% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pol2Ser2Reads.R
\name{Pol2Ser2Reads}
\alias{Pol2Ser2Reads}
\title{Pol2Ser2Reads}
\description{
calculates rpkm of pol2-ser2 binding from bigwig data, omitting first exons
}
\examples{
  psr <- Pol2Ser2Reads$new(geneSymbols="GATA2")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Pol2Ser2Reads-new}{\code{Pol2Ser2Reads$new()}}
\item \href{#method-Pol2Ser2Reads-getGeneSymbol}{\code{Pol2Ser2Reads$getGeneSymbol()}}
\item \href{#method-Pol2Ser2Reads-getGeneID}{\code{Pol2Ser2Reads$getGeneID()}}
\item \href{#method-Pol2Ser2Reads-getFirstExons}{\code{Pol2Ser2Reads$getFirstExons()}}
\item \href{#method-Pol2Ser2Reads-clone}{\code{Pol2Ser2Reads$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pol2Ser2Reads-new"></a>}}
\if{latex}{\out{\hypertarget{method-Pol2Ser2Reads-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pol2Ser2Reads$new(geneSymbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneSymbol}}{character, an indentifier for this object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of Pol2Ser2Reads
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pol2Ser2Reads-getGeneSymbol"></a>}}
\if{latex}{\out{\hypertarget{method-Pol2Ser2Reads-getGeneSymbol}{}}}
\subsection{Method \code{getGeneSymbol()}}{
accessor for the object's geneSybmol field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pol2Ser2Reads$getGeneSymbol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneSymbol
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pol2Ser2Reads-getGeneID"></a>}}
\if{latex}{\out{\hypertarget{method-Pol2Ser2Reads-getGeneID}{}}}
\subsection{Method \code{getGeneID()}}{
accessor for the object's entrez geneID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pol2Ser2Reads$getGeneID()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pol2Ser2Reads-getFirstExons"></a>}}
\if{latex}{\out{\hypertarget{method-Pol2Ser2Reads-getFirstExons}{}}}
\subsection{Method \code{getFirstExons()}}{
the firstexons and their range
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pol2Ser2Reads$getFirstExons()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pol2Ser2Reads-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Pol2Ser2Reads-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pol2Ser2Reads$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
