% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolymeraseReads.R
\name{PolymeraseReads}
\alias{PolymeraseReads}
\title{PolymeraseReads}
\description{
calculates rpkm of pol2-ser2 binding from bigwig data, omitting first exons
}
\examples{
  psr <- PolymeraseReads$new(geneSymbols="GATA2")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PolymeraseReads-new}{\code{PolymeraseReads$new()}}
\item \href{#method-PolymeraseReads-getGeneSymbol}{\code{PolymeraseReads$getGeneSymbol()}}
\item \href{#method-PolymeraseReads-getGeneID}{\code{PolymeraseReads$getGeneID()}}
\item \href{#method-PolymeraseReads-getFirstExons}{\code{PolymeraseReads$getFirstExons()}}
\item \href{#method-PolymeraseReads-clone}{\code{PolymeraseReads$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-new"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$new(geneSymbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneSymbol}}{character, an indentifier for this object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of PolymeraseReads
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-getGeneSymbol"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-getGeneSymbol}{}}}
\subsection{Method \code{getGeneSymbol()}}{
accessor for the object's geneSybmol field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$getGeneSymbol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneSymbol
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-getGeneID"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-getGeneID}{}}}
\subsection{Method \code{getGeneID()}}{
accessor for the object's entrez geneID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$getGeneID()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-getFirstExons"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-getFirstExons}{}}}
\subsection{Method \code{getFirstExons()}}{
the firstexons and their range
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$getFirstExons()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
