% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolymeraseReads.R
\name{PolymeraseReads}
\alias{PolymeraseReads}
\title{PolymeraseReads}
\description{
calculates rpkm of pol2-ser2 binding from bigwig data, omitting first exons
}
\examples{
  psr <- PolymeraseReads$new(geneSymbols="GATA2")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PolymeraseReads-new}{\code{PolymeraseReads$new()}}
\item \href{#method-PolymeraseReads-getGeneSymbol}{\code{PolymeraseReads$getGeneSymbol()}}
\item \href{#method-PolymeraseReads-getGeneID}{\code{PolymeraseReads$getGeneID()}}
\item \href{#method-PolymeraseReads-getTranscriptCoordinates}{\code{PolymeraseReads$getTranscriptCoordinates()}}
\item \href{#method-PolymeraseReads-getReads}{\code{PolymeraseReads$getReads()}}
\item \href{#method-PolymeraseReads-rpkm}{\code{PolymeraseReads$rpkm()}}
\item \href{#method-PolymeraseReads-clone}{\code{PolymeraseReads$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-new"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$new(geneSymbol, pol2.bigwigFile, rna.bigwigFile = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneSymbol}}{character, an indentifier for this object}

\item{\code{pol2.bigwigFile}}{character, full path to the aligned reads file}

\item{\code{rna.bigwigFile}}{character, full path to the aligned reads file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of PolymeraseReads
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-getGeneSymbol"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-getGeneSymbol}{}}}
\subsection{Method \code{getGeneSymbol()}}{
accessor for the object's geneSybmol field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$getGeneSymbol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneSymbol
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-getGeneID"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-getGeneID}{}}}
\subsection{Method \code{getGeneID()}}{
accessor for the object's entrez geneID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$getGeneID()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the geneID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-getTranscriptCoordinates"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-getTranscriptCoordinates}{}}}
\subsection{Method \code{getTranscriptCoordinates()}}{
the coordinates of the longest transcript
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$getTranscriptCoordinates()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-getReads"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-getReads}{}}}
\subsection{Method \code{getReads()}}{
the GRanges of scored and aligned reads of
the longest transcript of geneSymbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$getReads()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GRanges
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-rpkm"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-rpkm}{}}}
\subsection{Method \code{rpkm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$rpkm(
  geneSymbol,
  score.threshold = 0,
  assay,
  start.site.avoidance = 500,
  igv = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneSymbol}}{character HUGO gene symbol name}

\item{\code{score.threshold}}{numeric ignore reads scoring less than this value}

\item{\code{assay}}{character either "rna" or "pol2"}

\item{\code{start.site.avoidance}}{numeric default 500bp, for pol2 only, permits
rpkm calculation to avoid distortion by}

\item{\code{igv}}{default NA, a reference to an igvR instance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the rpkm score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolymeraseReads-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PolymeraseReads-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolymeraseReads$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
